Updated babel and jest and re-organized where the tests are located.
