+++ package.json
@@ -51,18 +51,16 @@
     "react-autocomplete": "^1.7.2"
   },
   "devDependencies": {
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.4.4",
     "@storybook/addon-actions": "^3.2.6",
     "@storybook/addon-knobs": "3.2.8",
     "@storybook/react": "^3.4.11",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
     "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.3",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-register": "^6.26.0",
+    "babel-jest": "^24.8.0",
     "enzyme": "^3.2.0",
     "enzyme-adapter-react-15": "^1.0.5",
     "eslint": "^4.14.0",
@@ -72,7 +70,7 @@
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-react": "^7.3.0",
-    "jest": "^20.0.4",
+    "jest": "^24.8.0",
     "jsdom": "^11.2.0",
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.25",
